break: remove `config` and `state`
dhall-rust is no longer maintained, and it is difficult to compile it to the Android platform. I decided to keep basic functionality first and configurability later.
