+++ liushu-core/Cargo.toml
@@ -10,7 +10,6 @@ serde = { version = "1", features = ["derive"] }
 csv = "1.1"
 directories = "4.0.1"
 once_cell = "1.17.1"
-serde_dhall = "0.12.1"
 pinyin = "0.9.0"
 redb = "0.20.0"
 regex = "1.7.1"
