+++ liushu-core/src/deploy.rs
@@ -1,36 +1,54 @@
-use std::fs::{self, File, OpenOptions};
+use std::fs::File;
 
-use crate::{config::Config, dirs::MyProjectDirs, engine::state::State, error::LiushuError};
+use patricia_tree::PatriciaMap;
+
+use crate::{
+    dict::{DictItem, DICTIONARY},
+    dirs::MyProjectDirs,
+    error::LiushuError,
+};
 
 pub fn deploy(proj_dirs: &MyProjectDirs) -> Result<(), LiushuError> {
-    let config = Config::load();
-    let mut state = State::from(&config);
+    let self_config_dir = &proj_dirs.config_dir.join("sunman");
+    let db_path = &proj_dirs.target_dir.join("sunman.redb");
 
-    let state_path = proj_dirs.data_dir.join(".state");
-    let old_state: State = File::open(&state_path)
-        .map_err(LiushuError::from)
-        .and_then(|r| bincode::deserialize_from(r).map_err(LiushuError::from))
-        .unwrap_or(State::default());
+    let table = redb::Database::create(db_path)?;
+    let tx = table.begin_write()?;
+    let mut trie = PatriciaMap::new();
+    {
+        let mut dict_table = tx.open_table(DICTIONARY)?;
+        let dictionaries = [
+            "words.dict.tsv",
+            "phrases.brief.dict.tsv",
+            "phrases.core.dict.tsv",
+        ];
+        for dict_path in dictionaries {
+            let dict_path = self_config_dir.join(dict_path);
+            let mut rdr = csv::ReaderBuilder::new()
+                .delimiter(b'\t')
+                .comment(Some(b'#'))
+                .from_path(dict_path)?;
+            for result in rdr.deserialize() {
+                let DictItem {
+                    text,
+                    code,
+                    weight,
+                    comment,
+                } = result?;
+                dict_table.insert(text.as_str(), (weight, comment.as_deref()))?;
 
-    fs::create_dir_all(&proj_dirs.target_dir)?;
-
-    for formula in &config.formulas {
-        let old_dict_digest = old_state
-            .get_avaliable_formula(&formula.id)
-            .map(|f| f.dict_digest.clone());
-        let new_dict_digest = formula.get_dictionaries_digest(&proj_dirs.config_dir);
-        if old_dict_digest != new_dict_digest {
-            formula.compile(proj_dirs)?;
-            state.set_dictionaries_digest(&formula.id, &new_dict_digest.unwrap());
+                if trie.get(&code).is_none() {
+                    trie.insert_str(code.as_str(), vec![text]);
+                } else if let Some(entry) = trie.get_mut(code.as_str()) {
+                    entry.push(text);
+                }
+            }
         }
     }
-
-    let state_writer = OpenOptions::new()
-        .write(true)
-        .truncate(true)
-        .create(true)
-        .open(&state_path)?;
-    bincode::serialize_into(state_writer, &state)?;
+    tx.commit()?;
+    let trie_path = &proj_dirs.target_dir.join("sunman.trie");
+    let trie_writer = File::create(trie_path)?;
+    bincode::serialize_into(trie_writer, &trie)?;
 
     Ok(())
 }
