+++ liushu-core/src/engine.rs
@@ -1,19 +1,15 @@
 pub mod candidates;
 mod segmentor;
-pub mod state;
 
-use std::{fs::File, path::PathBuf};
+use std::fs::File;
 
 use itertools::Itertools;
 use patricia_tree::PatriciaMap;
 use redb::{Database, ReadableTable};
 
-use crate::{dict::DICTIONARY, dirs::MyProjectDirs, error::LiushuError, hmm::pinyin_to_sentence};
+use crate::{dict::DICTIONARY, dirs::MyProjectDirs, error::LiushuError};
 
-use self::{
-    candidates::{Candidate, CandidateSource},
-    state::State,
-};
+use self::candidates::{Candidate, CandidateSource};
 
 pub trait InputMethodEngine {
     fn search(&self, code: &str) -> Result<Vec<Candidate>, LiushuError>;
@@ -21,53 +17,17 @@ pub trait InputMethodEngine {
 
 #[derive(Debug)]
 pub struct Engine {
-    target_dir: PathBuf,
-    state: State,
     db: Database,
     trie: PatriciaMap<Vec<String>>,
 }
 
 impl Engine {
     pub fn init(proj_dirs: &MyProjectDirs) -> Result<Self, LiushuError> {
-        let state: State =
-            bincode::deserialize_from(File::open(proj_dirs.data_dir.join(".state"))?)?;
+        let db = Database::open(proj_dirs.target_dir.join("sunman.redb"))?;
+        let trie: PatriciaMap<Vec<String>> =
+            bincode::deserialize_from(File::open(proj_dirs.target_dir.join("sunman.trie"))?)?;
 
-        let active_formula = state.get_active_formula();
-        let db = Database::open(
-            proj_dirs
-                .target_dir
-                .join(format!("{}.redb", active_formula.id)),
-        )?;
-        let trie: PatriciaMap<Vec<String>> = bincode::deserialize_from(File::open(
-            proj_dirs
-                .target_dir
-                .join(format!("{}.trie", active_formula.id)),
-        )?)?;
-
-        Ok(Self {
-            target_dir: proj_dirs.target_dir.clone(),
-            state,
-            db,
-            trie,
-        })
-    }
-
-    pub fn set_active_formula(&mut self, formula_id: &str) -> Result<(), LiushuError> {
-        if formula_id == self.state.get_active_formula_id() {
-            return Ok(());
-        }
-
-        self.state.set_active_formula(formula_id);
-        let active_formula = self.state.get_active_formula();
-        let db = Database::open(self.target_dir.join(format!("{}.redb", active_formula.id)))?;
-        let trie: PatriciaMap<Vec<String>> = bincode::deserialize_from(File::open(
-            self.target_dir.join(format!("{}.trie", active_formula.id)),
-        )?)?;
-
-        self.db = db;
-        self.trie = trie;
-
-        Ok(())
+        Ok(Self { db, trie })
     }
 }
 
@@ -76,7 +36,7 @@ impl InputMethodEngine for Engine {
         let tx = self.db.begin_read()?;
         let dictionary = tx.open_table(DICTIONARY)?;
 
-        let mut result: Vec<Candidate> = self
+        let result: Vec<Candidate> = self
             .trie
             .iter_prefix(code.as_bytes())
             .flat_map(|(key, value)| {
@@ -101,21 +61,6 @@ impl InputMethodEngine for Engine {
             .sorted_by_key(|i| std::cmp::Reverse(i.weight))
             .collect();
 
-        let active_formula = self.state.get_active_formula();
-        if active_formula.use_hmm && code.len() > 6 {
-            let pinyin_sequence = segmentor::split_pinyin(code, &self.trie);
-            let predict = pinyin_to_sentence(&pinyin_sequence, &self.db, &self.trie)?;
-            result.insert(
-                0,
-                Candidate {
-                    code: code.to_string(),
-                    text: predict,
-                    weight: u64::MAX,
-                    comment: None,
-                    source: CandidateSource::Hmm,
-                },
-            );
-        }
         Ok(result)
     }
 }
