+++ liushu-core/src/lib.rs
@@ -1,4 +1,3 @@
-mod config;
 pub mod deploy;
 pub mod dict;
 pub mod dirs;
